requirements.txt
openai>=1.6.1
langchain>=0.0.350
langchain-openai>=0.0.2
chromadb>=0.4.18
sentence-transformers>=2.2.2
streamlit>=1.28.1
python-dotenv>=1.0.0
tiktoken>=0.5.1
numpy>=1.24.3
pandas>=2.0.3 